% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_pixels.R
\name{classify_pixels}
\alias{classify_pixels}
\title{Classifies the pixels of an image}
\usage{
classify_pixels(image_prop, ..., unclassed_colour = "black", verbose = TRUE)
}
\arguments{
\item{image_prop}{an array containing the transformed colours. It must be an
object produced with function \code{transform_colours()}.}

\item{\dots}{a list of objects of class \code{pixel_cat} containing the
classification rules.}

\item{unclassed_colour}{a character string setting the colour to be assigned
to unclassified pixels. Defaults to "black".}

\item{verbose}{a logical value. When TRUE (default) the function prints some
statistics about the classification.}
}
\value{
An object of class \code{classified_image}, which is a list of lists.
  Each first-level element corresponds to one of the pixel categories and its
  name is the category name. They contains the second-level list, which have
  the following elements:
\itemize{
  \item \code{colour}: a string defining a colour to paint the pixels in the
  classified image. Inherited from the
  \code{pixel_class} object defining the class.
  \item \code{incid_mat}: a logical matrix where \code{TRUE} values indicate
  that the pixel belongs to this pixel category.
}
}
\description{
Classifies the pixels represented in an object of class
\code{transformed_image} using the rules contained in a list of objects of
class \code{pixel_cat}.
}
\details{
This function generates a set of incid matrices indicating
whether a pixel belongs to a pixel category or not. An additional matrix is
added identifying the pixels that do not belong to the defined categories.
}
\examples{
\dontrun{

# The series of steps to classify a fictitious image

yellow <- "#ffcd0eff"
blue <- "#5536ffff"

ivy_oak_rgb <- read_image(system.file("extdata", "IvyOak400x300.JPG",
                          package = "pixelclasser"))

rule_01 <- define_rule("rule_01", "g", "b", c(0.345, 1/3), c(0.40, 0.10),
                        comparator = "<")
rule_02 <- define_rule("rule_02", "g", "b", c(0.345, 1/3), c(0.40, 0.10),
                        comparator = ">=")

cat_dead_leaves <- define_cat("dead_leaves", blue, rule_01)
cat_living_leaves <- define_cat("living_leaves", yellow, rule_02)

dead_live_classified <- classify_pixels(ivy_oak_rgb, cat_dead_leaves,
                        cat_living_leaves)

}

}
\seealso{
\code{\link{define_cat}}, \code{\link[grDevices]{col2rgb}},
\code{\link[grDevices]{colours}}
}
