% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_subcat.R
\name{define_subcat}
\alias{define_subcat}
\title{Creates a subcategory object}
\usage{
define_subcat(subcat_name, ...)
}
\arguments{
\item{subcat_name}{a character string containing the name of the subcategory.}

\item{\dots}{a list of objects of class \code{pixel_rule}.}
}
\value{
An object of class \code{pixel_subcat}, which is a list with the
  following elements: \itemize{ \item \code{name} a character string
  containing the name of the subcategory. \item \code{rules_list} a list
  containing the rules. }
}
\description{
Creates a list of class \code{pixel_subcat} object from a list of objects of
class \code{pixel_rule}.
}
\examples{
rule01 <- define_rule("R01", "g", "b",
                      list(c(0.35, 0.30), c(0.45, 0.10)), ">=")
rule02 <- define_rule("R02", "g", "b",
                      list(c(0.35, 0.253), c(0.45, 0.253)), ">=")

subcat01 <- define_subcat("subcat01", rule01, rule02)

}
\seealso{
\code{\link{define_rule}}, \code{\link{define_cat}}
}
