% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{define_cat}
\alias{define_subcat}
\alias{define_rule}
\title{Deprecated functions}
\usage{
define_cat(cat_name, cat_colour, ...)

define_subcat(subcat_name, ...)

define_rule(rule_name, x_axis, y_axis, rule_points, comp_op)
}
\arguments{
\item{cat_name}{a character string containing the name of the category.}

\item{cat_colour}{a character string defining the colour to paint the pixels
with when creating a classified picture.}

\item{\dots}{a list of objects of class \code{pixel_rule}.}

\item{subcat_name}{a character string containing the name of the subcategory.}

\item{rule_name}{a character string containing the name of the rule.}

\item{x_axis}{a character string selecting the colour variable used as x
axis, one of \code{"r"}, \code{"g"} or \code{"b"}.}

\item{y_axis}{a character string selecting the colour variable used as y
axis, one of \code{"r"}, \code{"g"} or \code{"b"}.}

\item{rule_points}{either an object of  of class \code{"rule_points"} created
with function \code{place_rule()}, or a list containing the coordinates of
two points defining the line.}

\item{comp_op}{a character string containing one of the comparison operators
\code{">", ">=", "<", "<="}.}
}
\description{
Deprecated functions
}
\details{
These functions were constructors for pixelclasser objects, but now
  they are substituted by constructors with the same name as the class, as is
  customary in R.
}
\seealso{
\code{\link{pixel_category}}, \code{\link{pixel_subcategory}} and
  \code{\link{pixel_rule}}.
}
