% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_rule.R
\name{place_rule}
\alias{place_rule}
\title{Places a line on the rgb plot}
\usage{
place_rule(line_type = "normal")
}
\arguments{
\item{line_type}{a character string indicating that the line is
\code{"vertical"} or \code{"horizontal"}.}
}
\value{
A list containing the following elements: \itemize{ \item
  \code{first_point}: coordinates of the start point of the line. \item
  \code{second_point}: coordinates of the end point of the line. }
}
\description{
A wrapper function for \code{graphics:locator} that facilitates the creation
of rules.
}
\details{
This function calls \code{graphics::locator} allowing to select two
  points, then plots the line joining these points and returns a list
  containing the coordinates of the points. The coordinates are rearranged in
  a way convenient to be passed to \code{define_rule()}.

  True horizontal and vertical lines are difficult to create by hand. In 
  these cases, specifying \code{"vertical"} or \code{"horizontal"} (partial
  match allowed, i e "h") will copy the appropriate coordinate value from the
  first point to the second. Note that this is done after \code{locator()}
  returns, so the plot will show the line as was created by the user, not the
  corrected one. Use \code{plot_rule()} to see this.
}
\examples{
\dontrun{
After creating a plot with the points, the lines are created with
line01 <- place_rule()        # A "normal" line
line02 <- place_rule("h")     # An horizontal line}

}
\seealso{
\code{\link[graphics]{locator}}, \code{\link{define_rule}},
  \code{\link{plot_rule}}, \code{\link{plot_rgb_plane}}
}
