% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rule.R
\name{plot_rule}
\alias{plot_rule}
\title{Plots the line that defines a rule}
\usage{
plot_rule(rule, label = "", ...)
}
\arguments{
\item{rule}{an object of class \code{rule} produced by \code{define_rule()}.}

\item{label}{a string to label the line. It is attached at the coordinates of
the second point used to define the line.}

\item{\dots}{additional graphical parameters passed to the underlying
\code{lines()} function, for example to define the line colour or dashing
style.}
}
\value{
The function does not return any value.
}
\description{
This function draws the line that defines a rule on the plot created by
\code{plot_rgb_plane()}.
}
\details{
The function uses the information stored in the rule object to plot
  the line. Therefore the line starts and ends at the points stored in the
  rule object. If these points were not convenient for plotting, the
  coordinates of other points on the line can be calculated using its
  equation (the rule object contains the parameters of the line) and the
  original coordinates replaced by these.

  Caution: The colour variables used to define the rules and those used to
  plot the triangular graph must be the same.

  To set the colour of the line, use any character string understood by
  \code{col2rgb()}.
}
\examples{
rule_01 <- define_rule("rule_01", "g", "b", c(0.345, 1/3),
                       c(0.40, 0.10), comparator = "<")
rule_02 <- define_rule("rule_02","g", "b", c(0.35, 0.30),
                       c(0.565, 0.10), comparator = "<")

plot_rgb_plane("g", "b")

# The first rule is represented as a green line without label
plot_rule(rule_01, col = "green")

# And the second in black with a label (shifted to the right)
plot_rule(rule_02, label = expression('L'[1]*''), lty = 1,
          col = 'black', adj = 0)

}
\seealso{
\code{\link{plot_rgb_plane}}, \code{\link{define_rule}},
  \code{\link[grDevices]{col2rgb}}
}
