% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_colours.R
\name{transform_colours}
\alias{transform_colours}
\title{Transforms RGB values into proportions (rgb values)}
\usage{
transform_colours(image_array)
}
\arguments{
\item{image_array}{an array of class \code{image_array} created by function
\code{read_image()}.}
}
\value{
Returns an array of class \code{transformed_image} containing the
  proportions of each colour variable in the pixels of the image. The third
  dimension of the array is named "bands" and its elements are labelled as
  "r", "g" and "b", respectively.
}
\description{
This function transforms an array of RGB absolute values into a similar array
containing the proportion of each band (= colour variable): r g and b.
}
\details{
The proportions are calculated as \code{r} = \code{R} / (\code{R + G
  + B}), and so on. It is used by function read_image().
}
\keyword{internal}
